+++ PowerLit/Lib/Lighting.hlsl
@@ -6,15 +6,15 @@
 #include "Shadows.hlsl"
 #include "GI.hlsl"
 
-half3 VertexLighting(half3 worldPos,half3 normal,bool isLightOn=false){
-    half3 c = (half3)0;
+float3 VertexLighting(float3 worldPos,float3 normal,bool isLightOn=false){
+    float3 c = (float3)0;
     #if defined(_ADDITIONAL_LIGHTS_VERTEX)
     // branch_if(isLightOn)
     {
         int count = GetAdditionalLightsCount();
         for(int i=0;i<count;i++){
             Light light = GetAdditionalLight(i,worldPos);
-            half3 lightColor = light.color * light.distanceAttenuation;
+            float3 lightColor = light.color * light.distanceAttenuation;
             c += LightingLambert(lightColor,light.direction,normal);
         }
     }
@@ -26,7 +26,7 @@ half3 VertexLighting(half3 worldPos,half3 normal,bool isLightOn=false){
 // Light 
 ***/
 
-Light GetMainLight(half4 shadowCoord,half3 worldPos,half4 shadowMask,bool isReceiveShadow){
+Light GetMainLight(float4 shadowCoord,float3 worldPos,float4 shadowMask,bool isReceiveShadow){
     Light light = (Light)0;
     light.direction = _MainLightPosition.xyz;
     light.color = _MainLightColor.rgb;
@@ -36,7 +36,7 @@ Light GetMainLight(half4 shadowCoord,half3 worldPos,half4 shadowMask,bool isRece
 }
 
 
-Light GetMainLight(SurfaceInputData data,half4 shadowMask){
+Light GetMainLight(SurfaceInputData data,float4 shadowMask){
     Light mainLight = GetMainLight(data.inputData.shadowCoord,data.inputData.positionWS,shadowMask,true);
     return mainLight;
 }
@@ -51,9 +51,9 @@ void OffsetLight(inout Light mainLight){
     #endif
 }
 
-void InitBRDFData(SurfaceInputData surfaceInputData,inout half alpha,out BRDFData brdfData){
+void InitBRDFData(SurfaceInputData surfaceInputData,inout float alpha,out BRDFData brdfData){
     SurfaceData surfaceData = surfaceInputData.surfaceData;
-    half oneMinusReflectivityMetallic = OneMinusReflectivityMetallic(surfaceData.metallic);
+    float oneMinusReflectivityMetallic = OneMinusReflectivityMetallic(surfaceData.metallic);
     
     brdfData = (BRDFData)0;
     // brdfData.albedo = surfaceData.albedo;
@@ -80,14 +80,14 @@ void InitBRDFData(SurfaceInputData surfaceInputData,inout half alpha,out BRDFDat
     Minimalist cook torrance
     r2/(d*d * lh*lh *(4r+2))
 ***/
-half3 CalcDirectSpecularTerm(half r/*roughness*/,half3 lightDir,half3 viewDir,half3 normal){
-    half3 h = SafeNormalize(lightDir + viewDir);
-    half nh = saturate(dot(normal,h));
-    half lh = saturate(dot(lightDir,h));
+float3 CalcDirectSpecularTerm(float r/*roughness*/,float3 lightDir,float3 viewDir,float3 normal){
+    float3 h = SafeNormalize(lightDir + viewDir);
+    float nh = saturate(dot(normal,h));
+    float lh = saturate(dot(lightDir,h));
 
-    half r2 = r * r;
-    half d = nh * nh * (r2-1)+1;
-    half specTerm = r2/( d * d * max(0.1, lh * lh) * ( 4 * r + 2 ));
+    float r2 = r * r;
+    float d = nh * nh * (r2-1)+1;
+    float specTerm = r2/( d * d * max(0.1, lh * lh) * ( 4 * r + 2 ));
 
     #if defined (SHADER_API_MOBILE) || defined (SHADER_API_SWITCH)
         specTerm = clamp(specTerm,0,100);
@@ -95,22 +95,22 @@ half3 CalcDirectSpecularTerm(half r/*roughness*/,half3 lightDir,half3 viewDir,ha
     return specTerm;
 }
 
-half3 CalcPBRLighting(BRDFData brdfData,half3 lightColor,half3 lightDir,half3 lightAtten,
-    half3 normal,half3 viewDir){
-    half nl = saturate(dot(normal,lightDir));
-    half3 radiance = lightColor * lightAtten * nl; // light's color
+float3 CalcPBRLighting(BRDFData brdfData,float3 lightColor,float3 lightDir,float3 lightAtten,
+    float3 normal,float3 viewDir){
+    float nl = saturate(dot(normal,lightDir));
+    float3 radiance = lightColor * lightAtten * nl; // light's color
 
-    half3 brdf = brdfData.diffuse;
+    float3 brdf = brdfData.diffuse;
     brdf += brdfData.specular * CalcDirectSpecularTerm(brdfData.roughness,lightDir,viewDir,normal);
     return brdf * radiance;
 }
 
-half3 CalcAdditionalPBRLighting(BRDFData brdfData,InputData inputData,half4 shadowMask){
+float3 CalcAdditionalPBRLighting(BRDFData brdfData,InputData inputData,float4 shadowMask){
     uint lightCount = GetAdditionalLightsCount();
-    half3 c = (half3)0;
+    float3 c = (float3)0;
     for(uint i=0;i<lightCount;i++){
         Light light = GetAdditionalLight1(i,inputData.positionWS,shadowMask);
-        // half3 attenColor = max(light.shadowAttenuation,inputData.bakedGI);
+        // float3 attenColor = max(light.shadowAttenuation,inputData.bakedGI);
         OffsetLight(light/**/);
 
         // branch_if(light.distanceAttenuation)
@@ -120,7 +120,7 @@ half3 CalcAdditionalPBRLighting(BRDFData brdfData,InputData inputData,half4 shad
 }
 
 
-half4 CalcPBR(SurfaceInputData data,Light mainLight,half4 shadowMask){
+float4 CalcPBR(SurfaceInputData data,Light mainLight,float4 shadowMask){
     SurfaceData surfaceData = data.surfaceData;
     InputData inputData = data.inputData;
 
@@ -129,8 +129,8 @@ half4 CalcPBR(SurfaceInputData data,Light mainLight,half4 shadowMask){
     
     // MixRealtimeAndBakedGI(mainLight,inputData.normalWS,inputData.bakedGI);
 // return (brdfData.diffuse + inputData.bakedGI*0.2).xyzx+shadowMask*0.1;
-    half customIBLMask = _IBLMaskMainTexA ? surfaceData.alpha : 1;
-    half3 color = CalcGI(brdfData,inputData.bakedGI,surfaceData.occlusion,inputData.normalWS,inputData.viewDirectionWS,customIBLMask,inputData.positionWS,data.screenUV);
+    float customIBLMask = _IBLMaskMainTexA ? surfaceData.alpha : 1;
+    float3 color = CalcGI(brdfData,inputData.bakedGI,surfaceData.occlusion,inputData.normalWS,inputData.viewDirectionWS,customIBLMask,inputData.positionWS,data.screenUV);
     // branch_if(mainLight.distanceAttenuation)
     {
         OffsetLight(mainLight/**/);
@@ -153,7 +153,7 @@ half4 CalcPBR(SurfaceInputData data,Light mainLight,half4 shadowMask){
     }
     #endif
 
-    return half4(color,surfaceData.alpha);
+    return float4(color,surfaceData.alpha);
 }
 
 #endif //LIGHTING_HLSL
